fix: predict_gpの標準化計算をfloatからdoubleに統一(e6ed9b9と同型)
e6ed9b9でpredict_linear/predict_mlpの標準化(x-mean)/scaleを倍精度化したが、
predict_gpだけ同じ標準化パターン(xs[i] = (x[i]-mean[i])/scale[i])がfloat演算の
まま見落とされていた。JS版(predict-core.js)は倍精度で計算するため、極端な入力値では
float32のULPレベルの丸め順序差がJS/C++間の誤差として現れる恐れがあった。他の
predict_*関数と同じ「倍精度で計算→最後に1回だけfloatへ丸める」パターンに統一。

matrix_cpp_out/のgp_*(roundFalse系)・blend_all_types_mixed_none_roundFalseの期待値を
MinGW g++ビルドで再生成(round_output=trueの設定は整数丸めで差が吸収されるため無変化)。
diffは1e-7オーダー(float32 ULP相当)のみで、npm testは50/50全PASSを維持。

+++ native_predictor/predict_native_v2.cpp
@@ -501,16 +501,21 @@ static float predict_lgbm(const LGBMModel& m, const float* x) {
 }
 
 static float predict_gp(const GPModel& m, const float* x) {
+    // e6ed9b9フォロー(重-predict_gp): predict_linear/predict_mlpはこの標準化を倍精度で
+    // 計算してから最後に1回だけfloatへ丸めるよう統一済みだが、predict_gpだけfloatの
+    // まま(xs[i] = (x[i]-mean[i])/scale[i] を float 演算)残っていた。極端な入力値では
+    // JS版(倍精度で計算)とfloat32丸め順序差が有意な誤差になる恐れがあるため、同じ
+    // パターンで倍精度化する。
     const int d = m.n_feat;
-    std::vector<float> xs(d);
+    std::vector<double> xs(d);
     for (int i = 0; i < d; i++)
-        xs[i] = (x[i] - m.mean[i]) / m.scale[i];
+        xs[i] = ((double)x[i] - (double)m.mean[i]) / (double)m.scale[i];
 
     double y_norm = 0.0;
     for (int i = 0; i < m.n_train; i++) {
         double sq = 0.0;
         for (int j = 0; j < d; j++) {
-            double diff = xs[j] - m.X_train[i * d + j];
+            double diff = xs[j] - (double)m.X_train[i * d + j];
             double ls_j = m.ls[j];
             sq += (diff / ls_j) * (diff / ls_j);
         }
